Add AnnotateRoutes::AnnotationProcessor
