+++ lib/annotate/annotate_routes.rb
@@ -19,17 +19,14 @@
 #
 
 require_relative './annotate_routes/helpers'
-require_relative './annotate_routes/header_generator'
+require_relative './annotate_routes/annotation_processor'
 
 module AnnotateRoutes
   class << self
     def do_annotations(options = {})
       if routes_file_exist?
         existing_text = File.read(routes_file)
-        content, header_position = Helpers.strip_annotations(existing_text)
-        new_content = annotate_routes(HeaderGenerator.generate(options), content, header_position, options)
-        new_text = new_content.join("\n")
-        if rewrite_contents(existing_text, new_text, options[:frozen])
+        if AnnotationProcessor.update(routes_file, existing_text, options)
           puts "#{routes_file} was annotated."
         else
           puts "#{routes_file} was not changed."
@@ -90,30 +87,6 @@ module AnnotateRoutes
       end
 
       content_changed
-    end
-
-    def annotate_routes(header, content, header_position, options = {})
-      magic_comments_map, content = Helpers.extract_magic_comments_from_array(content)
-      if %w(before top).include?(options[:position_in_routes])
-        header = header << '' if content.first != ''
-        magic_comments_map << '' if magic_comments_map.any?
-        new_content = magic_comments_map + header + content
-      else
-        # Ensure we have adequate trailing newlines at the end of the file to
-        # ensure a blank line separating the content from the annotation.
-        content << '' unless content.last == ''
-
-        # We're moving something from the top of the file to the bottom, so ditch
-        # the spacer we put in the first time around.
-        content.shift if header_position == :before && content.first == ''
-
-        new_content = magic_comments_map + content + header
-      end
-
-      # Make sure we end on a trailing newline.
-      new_content << '' unless new_content.last == ''
-
-      new_content
     end
   end
 end
